Move master specific code to own file
